Enforcing regex to be string literal
Previously regex expression for `valid` could have been anything string.
This a security and bug implications. This changeset moves away from
the notion of string to only allow literals and since the only level to
detect literals at JMES parser level is `&` (expression reference) so we
are going to use the expression references to ensure string literals.
The syntax IMO does not look really bad but could have been better.